+++ internal/msix/package_other.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package msix
+
+func isPackaged() (bool, error) {
+	return false, nil
+}